+++ acoustic_agent/__init__.py
@@ -3,7 +3,7 @@
 __version__ = "0.1.0"
 
 from .api import AcousticAgent, DynamicSimulationResult, MultiSourceSimulationResult
-from .audio import mix_audio, render_audio, resample_audio
+from .audio import mix_audio, mix_audio_at_snr, render_audio, resample_audio
 from .batch import BatchResult, ProductionResult, SimulationPair, simulate_batch
 from .custom_floorplan import FloorplanBuilder
 from .directivity import source_directivity
@@ -39,6 +39,7 @@ __all__ = [
     "SimulationResult",
     "make_room",
     "mix_audio",
+    "mix_audio_at_snr",
     "microphone_array",
     "simulate_batch",
     "simulate_rir",